+++ 5.1/5.1.6/alpine/docker-initialize.py
@@ -148,6 +148,11 @@ class Environment(object):
 
         sources = self.env.get("SOURCES", "").strip().split(",")
 
+        buildout_extends = ((develop or sources)
+                            and ["develop.cfg"] or ["buildout.cfg"])
+        extra_extends = self.env.get("BUILDOUT_EXTENDS", "").strip().split()
+        buildout_extends.extend(extra_extends)
+
         # If profiles not provided. Install ADDONS :default profiles
         if not profiles:
             for egg in eggs:
@@ -155,10 +160,11 @@ class Environment(object):
                 profiles.append("%s:default" % base)
 
         enabled = bool(site)
-        if not (eggs or zcml or develop or enabled):
+        if not (eggs or zcml or develop or enabled or extra_extends):
             return
 
         buildout = BUILDOUT_TEMPLATE.format(
+            buildout_extends="\n\t".join(buildout_extends),
             findlinks="\n\t".join(findlinks),
             eggs="\n\t".join(eggs),
             zcml="\n\t".join(zcml),
@@ -223,7 +229,7 @@ CORS_TEMPLATE = """<configure
 
 BUILDOUT_TEMPLATE = """
 [buildout]
-extends = develop.cfg
+extends = {buildout_extends}
 find-links += {findlinks}
 develop += {develop}
 eggs += {eggs}